Fix error message while using file picker
b/16092826 OrbitGgp/Client.cpp: Checks now when the timeout is triggered
whether a the a ggp request actually already finished. This prevents an
error message to popup when the "Load Capture" file picker is open while
the list of instances is loaded.

+++ OrbitGgp/Client.cpp
@@ -31,12 +31,14 @@ void RunProcessWithTimeout(
 
   QObject::connect(timeout_timer, &QTimer::timeout, parent,
                    [process, timeout_timer, callback]() {
-                     ERROR("Process request timed out after %dms",
-                           kDefaultTimeoutInMs);
-                     callback(Error::kRequestTimedOut);
-                     process->terminate();
-                     process->waitForFinished();
-                     if (process != nullptr) process->deleteLater();
+                     if (!process->waitForFinished(10)) {
+                       ERROR("Process request timed out after %dms",
+                             kDefaultTimeoutInMs);
+                       callback(Error::kRequestTimedOut);
+                       process->terminate();
+                       process->waitForFinished();
+                       process->deleteLater();
+                     }
 
                      timeout_timer->deleteLater();
                    });
